Add tests for Kafka and MQTT publishers
- Implement additional tests for Kafka producer message carrier, including span creation and topic attributes.
- Enhance MQTT publisher tests to ensure publishing properties and observability headers are correctly injected.
- Validate error handling in the MQTT publisher configuration.
- Ensure existing properties are preserved when injecting observability headers in MQTT messages.
